test(logout): move handler regression tests to integration test
- Moved logout_handlers regression tests from lib.rs unit tests to
  src-tauri/tests/logout_handlers.rs integration test.
- Integration test uses include_str!(../src/lib.rs) so the searched
  source does NOT contain the assertions themselves, preventing false
  positives where the test passes because the string appears inside the
  test code (Copilot review feedback, round 9).
- All tests pass: 72 unit + 2 integration = 74 total.
